Migrate sveltekit breaking change
