+++ package.json
@@ -6,7 +6,7 @@
     "build": "vite build",
     "package": "svelte-kit package",
     "preview": "vite preview",
-    "prepare": "svelte-kit sync && husky install",
+    "prepare": "husky install",
     "test": "playwright test",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
@@ -15,10 +15,9 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.22.2",
-    "@sveltejs/adapter-static": "^1.0.0-next.38",
-    "@sveltejs/kit": "1.0.0-next.385",
+    "@sveltejs/adapter-static": "^1.0.0-next.42",
+    "@sveltejs/kit": "^1.0.0-next.480",
     "@types/big.js": "^6.1.5",
-    "@types/crypto-js": "^4.1.1",
     "@types/hammerjs": "^2.0.41",
     "@types/lodash.debounce": "^4.0.7",
     "@types/lodash.isequal": "^4.5.6",
@@ -35,13 +34,13 @@
     "postcss": "^8.4.14",
     "prettier": "^2.6.2",
     "prettier-plugin-svelte": "^2.7.0",
-    "svelte": "^3.44.0",
-    "svelte-check": "^2.7.1",
-    "svelte-preprocess": "^4.10.6",
+    "svelte": "^3.50.1",
+    "svelte-check": "^2.9.0",
+    "svelte-preprocess": "^4.10.7",
     "tailwindcss": "^3.1.6",
-    "tslib": "^2.3.1",
-    "typescript": "^4.8.2",
-    "vite": "^3.0.1"
+    "tslib": "^2.4.0",
+    "typescript": "^4.8.3",
+    "vite": "^3.1.0"
   },
   "type": "module",
   "dependencies": {
