--- src/hooks/index.ts
@@ -1,5 +1,0 @@
-import type { Handle } from '@sveltejs/kit'
-
-export const handle: Handle = ({ event, resolve }) => {
-  return resolve(event, { ssr: false })
-}
