+++ src/routes/+layout.ts
@@ -1,0 +1,15 @@
+import type { LayoutLoad } from './$types'
+import { locale, loadTranslations } from '$lib/i18n/translations'
+
+export const load: LayoutLoad = async ({ url }) => {
+  const { pathname } = url
+  const defaultLocale = 'en'
+  const initLocale = locale.get() || defaultLocale
+
+  await loadTranslations(initLocale, pathname)
+
+  return {}
+}
+
+export const prerender = false
+export const ssr = false
