+++ src/routes/payments/[id]/+page.svelte
@@ -1,30 +1,15 @@
-<script lang="ts" context="module">
-  import type { Load } from './__types/[id]'
-
-  export const load: Load = ({ params }) => {
-    if (!credentials$.getValue().connection) {
-      return {
-        redirect: '/welcome',
-        status: 302
-      }
-    } else {
-      return { props: { id: params.id } }
-    }
-  }
-</script>
-
 <script lang="ts">
   import { fade } from 'svelte/transition'
-  import { credentials$, lastPath$, payments$ } from '$lib/streams'
+  import { lastPath$, payments$ } from '$lib/streams'
   import { goto } from '$app/navigation'
   import PaymentDetails from '$lib/components/PaymentDetails.svelte'
   import BackButton from '$lib/elements/BackButton.svelte'
   import { translate } from '$lib/i18n/translations'
   import Spinner from '$lib/elements/Spinner.svelte'
 
-  export let id: string // payment id
+  export let data: { id: string } // payment id
 
-  $: payment = $payments$.data && $payments$.data.find((p) => p.id === id)
+  $: payment = $payments$.data && $payments$.data.find((p) => p.id === data.id)
 
   function handleClose() {
     const path = lastPath$.value
