+++ src/routes/send/+page.ts
@@ -1,0 +1,1 @@
+export { load } from '$lib/utils'
