+++ src/routes/settings/app/unit/+page.svelte
@@ -1,7 +1,3 @@
-<script lang="ts" context="module">
-  export { load } from '$lib/utils'
-</script>
-
 <script lang="ts">
   import { goto } from '$app/navigation'
   import { fade } from 'svelte/transition'
