+++ src/routes/settings/app/unit/+page.ts
@@ -1,0 +1,1 @@
+export { load } from '$lib/utils'
