+++ src/routes/settings/help/+page.ts
@@ -1,0 +1,1 @@
+export { load } from '$lib/utils'
