+++ svelte.config.js
@@ -12,7 +12,7 @@ const config = {
       fallback: '200.html'
     }),
     prerender: {
-      default: false
+      entries: []
     }
   }
 }
